+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Disable fireplaces during day to reduce fuel consumption (except campfire and hearth).
+
 ## [0.6.26] - 2024-06-19
 
 ### Changed